[Notepad App] Added Italic and Underline Functionality
This commit implements the functionality for the "Italic" and
"Underline" buttons in the notepad app. The following updates have been
made:

1)Added a push button for "Italic" formatting.
2)Clicking the "Italic" button now toggles the italic formatting of the
selected text in the notepad.
3)Added a push button for "Underline" formatting.
4)Clicking the "Underline" button now toggles the underline formatting
of the selected text in the notepad.

These additions enhance the editing capabilities of the notepad app by
providing the ability to apply italic and underline formatting to the
text.